refactor: move translations from app dir to root dir
