+++ app/lib/i18n/translation.js
@@ -1,23 +1,23 @@
 // Load translation data.
-import bgTranslationMessages from '../../translations/bg-BG.json'
-import csTranslationMessages from '../../translations/cs-CZ.json'
-import deTranslationMessages from '../../translations/de-DE.json'
-import elTranslationMessages from '../../translations/el-GR.json'
-import enTranslationMessages from '../../translations/en.json'
-import esTranslationMessages from '../../translations/es-ES.json'
-import frTranslationMessages from '../../translations/fr-FR.json'
-import gaTranslationMessages from '../../translations/ga-IE.json'
-import hrTranslationMessages from '../../translations/hr-HR.json'
-import jaTranslationMessages from '../../translations/ja-JP.json'
-import nlTranslationMessages from '../../translations/nl-NL.json'
-import ptTranslationMessages from '../../translations/pt-BR.json'
-import roTranslationMessages from '../../translations/ro-RO.json'
-import ruTranslationMessages from '../../translations/ru-RU.json'
-import svTranslationMessages from '../../translations/sv-SE.json'
-import trTranslationMessages from '../../translations/tr-TR.json'
-import ukTranslationMessages from '../../translations/uk-UA.json'
-import zhCNTranslationMessages from '../../translations/zh-CN.json'
-import zhTWTranslationMessages from '../../translations/zh-TW.json'
+import bgTranslationMessages from '../../../translations/bg-BG.json'
+import csTranslationMessages from '../../../translations/cs-CZ.json'
+import deTranslationMessages from '../../../translations/de-DE.json'
+import elTranslationMessages from '../../../translations/el-GR.json'
+import enTranslationMessages from '../../../translations/en.json'
+import esTranslationMessages from '../../../translations/es-ES.json'
+import frTranslationMessages from '../../../translations/fr-FR.json'
+import gaTranslationMessages from '../../../translations/ga-IE.json'
+import hrTranslationMessages from '../../../translations/hr-HR.json'
+import jaTranslationMessages from '../../../translations/ja-JP.json'
+import nlTranslationMessages from '../../../translations/nl-NL.json'
+import ptTranslationMessages from '../../../translations/pt-BR.json'
+import roTranslationMessages from '../../../translations/ro-RO.json'
+import ruTranslationMessages from '../../../translations/ru-RU.json'
+import svTranslationMessages from '../../../translations/sv-SE.json'
+import trTranslationMessages from '../../../translations/tr-TR.json'
+import ukTranslationMessages from '../../../translations/uk-UA.json'
+import zhCNTranslationMessages from '../../../translations/zh-CN.json'
+import zhTWTranslationMessages from '../../../translations/zh-TW.json'
 
 // Collate all translations.
 const translations = {
