+++ crowdin.yml
@@ -1,5 +1,5 @@
 commit_message: '[ci skip]'
 files:
-  - source: /app/translations/en.json
-    translation: /app/translations/%locale%.json
+  - source: /translations/en.json
+    translation: /translations/%locale%.json
     update_option: update_without_changes
